Add files via upload
It is necessary to model the object of type Book with the following characteristics: title, price, stock, and Author. The Author has the characteristics of name, last name, email, and gender ('M' or 'F'). For this exercise, we will assume that a book has a single author. Execute the following instructions:
a. Initialize an object of type Author, "Joshua", "Bloch", "joshua@email.com", 'M'.
b. Print the previously instantiated author to the screen.
c. Initialize the book "Effective Java" by Author "Joshua Bloch" which costs 450 pesos with a quantity of 150 copies.
d. Print the instantiated book to the screen.
e. Modify the price of the book "Effective Java" to 500 pesos and increase the quantity by 50 copies.
f. Print the attributes of the author Joshua, accessed from the book "Effective Java".
g. Add a method to the Book class that enables printing the following message to the screen: "The book, {title} by {author's name}. It is sold at {price} pesos."